Move utils from API to SDK.
