+++ sdk/src/test/java/io/opentelemetry/sdk/trace/util/LinksTest.java
@@ -14,13 +14,17 @@
  * limitations under the License.
  */
 
-package io.opentelemetry.trace.util;
+package io.opentelemetry.sdk.trace.util;
 
 import static com.google.common.truth.Truth.assertThat;
 
 import com.google.common.testing.EqualsTester;
+import io.opentelemetry.sdk.trace.TestUtils;
 import io.opentelemetry.trace.AttributeValue;
-import io.opentelemetry.trace.Event;
+import io.opentelemetry.trace.Link;
+import io.opentelemetry.trace.SpanContext;
+import io.opentelemetry.trace.TraceFlags;
+import io.opentelemetry.trace.Tracestate;
 import java.util.HashMap;
 import java.util.Map;
 import org.junit.Before;
@@ -28,10 +32,17 @@ import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.junit.runners.JUnit4;
 
-/** Unit tests for {@link Events}. */
+/** Unit tests for {@link Links}. */
 @RunWith(JUnit4.class)
-public class EventsTest {
+public class LinksTest {
   private final Map<String, AttributeValue> attributesMap = new HashMap<>();
+  private final Tracestate tracestate = Tracestate.builder().build();
+  private final SpanContext spanContext =
+      SpanContext.create(
+          TestUtils.generateRandomTraceId(),
+          TestUtils.generateRandomSpanId(),
+          TraceFlags.getDefault(),
+          tracestate);
 
   @Before
   public void setUp() {
@@ -42,32 +53,33 @@ public class EventsTest {
 
   @Test
   public void create() {
-    Event event = Events.create("test", attributesMap);
-    assertThat(event.getName()).isEqualTo("test");
-    assertThat(event.getAttributes()).isEqualTo(attributesMap);
+    Link link = Links.create(spanContext, attributesMap);
+    assertThat(link.getContext()).isEqualTo(spanContext);
+    assertThat(link.getAttributes()).isEqualTo(attributesMap);
   }
 
   @Test
   public void create_NoAttributes() {
-    Event event = Events.create("test");
-    assertThat(event.getName()).isEqualTo("test");
-    assertThat(event.getAttributes()).isEmpty();
+    Link link = Links.create(spanContext);
+    assertThat(link.getContext()).isEqualTo(spanContext);
+    assertThat(link.getAttributes()).isEmpty();
   }
 
   @Test
   public void link_EqualsAndHashCode() {
     EqualsTester tester = new EqualsTester();
     tester
-        .addEqualityGroup(Events.create("test"), Events.create("test"))
+        .addEqualityGroup(Links.create(spanContext), Links.create(spanContext))
         .addEqualityGroup(
-            Events.create("test", attributesMap), Events.create("test", attributesMap));
+            Links.create(spanContext, attributesMap), Links.create(spanContext, attributesMap));
     tester.testEquals();
   }
 
   @Test
   public void link_ToString() {
-    Event event = Events.create("test", attributesMap);
-    assertThat(event.toString()).contains("test");
-    assertThat(event.toString()).contains(attributesMap.toString());
+    Link link = Links.create(spanContext, attributesMap);
+    assertThat(link.toString()).contains(spanContext.getTraceId().toString());
+    assertThat(link.toString()).contains(spanContext.getSpanId().toString());
+    assertThat(link.toString()).contains(attributesMap.toString());
   }
 }
